Refactored for clarity
- removed ``spawner`` and ``stacker`` packages and moved associated files to ``support`` package.
- changed ``@lombok.Getter`` to ``@Getter`` and imported lombok annotations
